backend/catalog-service/requirements.txt
fastapi[standard]
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
redis==5.0.1
pytest
pyjwt
pytest-asyncio
httpx
requests
prometheus-fastapi-instrumentator
